requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
SQLAlchemy==2.0.44
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.7
python-dotenv==1.0.1
